==> 04_ocr_bert/01_img_to_txt.ipynb <==
# 01_ Import modules

import pytesseract

# 02_ Create pipeline

import os
import pytesseract

img_path = '/home/e077926/buscode_2023/05_E2_tata/data/SROIE2019/train/img'

for img in os.listdir(img_path):
    # Vérifier si le fichier est une image (ajuster les extensions selon vos besoins)
    if img.lower().endswith(('.png', '.jpg', '.jpeg')):
        img_full_path = os.path.join(img_path, img)
        img_name, _ = os.path.splitext(os.path.basename(img_full_path))

        text = pytesseract.image_to_string(img_full_path, lang='eng')
        
        with open(f'/home/e077926/buscode_2023/05_E2_tata/data/SROIE2019/train/text/{img_name}.txt', 'w') as f:
            f.write(text)


# img_path1 = '/home/e077926/buscode_2023/05_E2_tata/data/SROIE2019/train/img/X00016469619.jpg'
# text = pytesseract.image_to_string(img_path1,lang='eng')
# print(text)
