==> 04_ocr_bert/02_create_base_dataset.ipynb <==
# 01 Create values df

import os
import pandas as pd
import re

txt_data_path = "/home/e077926/buscode_2023/05_E2_tata/data/SROIE2019/train/text"
df_list = []

for txt_file in os.listdir(txt_data_path):
    # Construire le chemin complet du fichier texte
    txt_file_path = os.path.join(txt_data_path, txt_file)

    # Extraire l'ID du document à partir du nom du fichier sans extension
    doc_id = os.path.splitext(txt_file)[0]

    # Lire les lignes du fichier texte
    with open(txt_file_path, 'r') as f:
        lines = f.readlines()
        # Vérifier si le fichier n'est pas vide et exclure les lignes qui ne contiennent que '\n'
        filtered_lines = [line.strip() for line in lines if line.strip()]

        # Vérifier si des lignes restent après le filtrage
        if filtered_lines:
            # Créer un DataFrame pour les lignes du fichier texte
            df_lines = pd.DataFrame({'text_lines': filtered_lines})

            # Ajouter les colonnes d'ID du document et d'ID de ligne
            df_lines['doc_id'] = doc_id
            df_lines['line_id'] = df_lines.index + 1  # Index + 1 pour obtenir l'ID de ligne

            # Ajouter ce DataFrame à la liste
            df_list.append(df_lines)

# Concaténer tous les DataFrames dans la liste en un seul DataFrame
df = pd.concat(df_list, ignore_index=True)

df.tail(10)

# 02 Clean data

df['text_lines'] = df['text_lines'].apply(lambda x: re.sub(r'[^a-zA-Z0-9\s]', '', x))
df['text_lines'] = df['text_lines'].str.lower()

df.tail(10)

df.to_csv('/home/e077926/buscode_2023/05_E2_tata/data/receipt_raw_df.csv')
