==> 04_ocr_bert/04_training.ipynb <==
https://github.com/karndeepsingh/Named-Entity-Recognition/blob/main/NAMED%20ENTITY%20RECOGNITION.ipynb

# 01_ Load modules

# !pip install simpletransformers

from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from simpletransformers.ner import NERModel,NERArgs
import pandas as pd 
import numpy as np

df=pd.read_csv('/home/e077926/buscode_2023/05_E2_tata/data/bio_df.csv')
df.drop(columns=['Unnamed: 0'], inplace=True)
df.head()

# 02_ encode doc and lines

df_doc_encod=df.copy()
# Créer une nouvelle colonne 'sentence' en combinant 'doc' et 'line_id'
df_doc_encod['sentence_id'] = df_doc_encod['doc'] + df_doc_encod['line_id'].astype(str)
df_doc_encod['words']=df_doc_encod['value']
df_doc_encod['labels']=df_doc_encod['entity']
df_doc_encod=df_doc_encod.drop(columns=['doc','line_id', 'value', 'entity'])
df_doc_encod["sentence_id"] = LabelEncoder().fit_transform(df_doc_encod["sentence_id"] )

df_doc_encod.tail(10)

df_doc_encod["sentence_id"] = LabelEncoder().fit_transform(df_doc_encod["sentence_id"] )

df_doc_encod.tail(10)

df_doc_encod['labels']=df_doc_encod['labels'].map({'O':'O',
                          'address-B': 'B-ADD',
                          'address-I': 'I-ADD',
                          'date-B': 'B-TIM','date-I': 'I-TIM',
                          'company-B': 'B-ORG',
                          'company-I': 'I-ORG',
                          'total-B':'B-TOT','total-I':'I-TOT'})

# df_doc_encod['sentence_id']=df_doc_encod['doc']+df_doc_encod['line_id']
# df_doc_encod=df_doc_encod.drop(columns=['doc', 'line_id'])
# df_doc_encod=df_doc_encod.sort_values(by=['sentence_id', 'value', 'entity'])
# df_doc_encod.tail(10)

ner_tags_tuple_list = [('O', 0),
            ('B-ORG', 1),
            ('I-ORG', 2),
            ('B-ADD', 3),
            ('I-ADD', 4),
            ('B-TIM', 5),
            ('I-TOT', 6),
            ('B-TOT', 7),
            ('I-TIM', 8)]

# Création du dictionnaire
tag2index = {key: value for key, value in ner_tags_tuple_list}
index2tag = {value: key for key, value in ner_tags_tuple_list}
index2tag

df_doc_encod['ner_tags']=df_doc_encod['labels'].map(tag2index)
df_doc_encod.head()

# 03_ Split data

X=df_doc_encod[["sentence_id", "words"]]
Y=df_doc_encod["ner_tags"]

x_train, x_test, y_train, y_test = train_test_split(X,Y, test_size =0.2)

#building up train data and test data
train_data = pd.DataFrame({"sentence_id":x_train["sentence_id"],"words":x_train["words"],"ner_tags":y_train})
test_data = pd.DataFrame({"sentence_id":x_test["sentence_id"],"words":x_test["words"],"ner_tags":y_test})

# 04 Model Training

labels= df_doc_encod["ner_tags"].unique().tolist()
labels

args = NERArgs()
args.num_train_epochs = 1
args.learning_rate = 1e-4
args.overwrite_output_dir =True
args.train_batch_size = 32
args.eval_batch_size = 32

model = NERModel('bert', 'bert-base-cased',labels=labels,args =args)

model.train_model(train_data,eval_data = test_data,acc=accuracy_score)

# Test

https://github.com/laxmimerit/NLP-Tutorials-with-HuggingFace/blob/main/2%20NER%20Training%20%7C%20NLP%20with%20HuggingFace%20Tutorial.ipynb

from transformers import AutoTokenizer

model_checkpoint = "distilbert-base-cased"
tokenizer = AutoTokenizer.from_pretrained(model_checkpoint)

ner_tags_tuple_list = [('O', 0),
            ('B-ORG', 1),
            ('I-ORG', 2),
            ('B-ADD', 3),
            ('I-ADD', 4),
            ('B-TIM', 5),
            ('I-TOT', 6),
            ('B-TOT', 7),
            ('I-TIM', 8)]

# Création du dictionnaire
tag2index = {key: value for key, value in ner_tags_tuple_list}
index2tag = {value: key for key, value in ner_tags_tuple_list}
index2tag

df_doc_encod['ner_tags']=df_doc_encod['labels'].map(tag2index)
df_doc_encod.head()

df_doc_encod.to_csv('/home/e077926/buscode_2023/05_E2_tata/data/final_dataset.csv')

from transformers import AutoModelForTokenClassification

model = AutoModelForTokenClassification.from_pretrained(
                                                    model_checkpoint,
                                                    id2label=index2tag,
                                                    label2id=tag2index)

from transformers import TrainingArguments

args = TrainingArguments("distilbert-finetuned-ner",
                         evaluation_strategy = "epoch",
                         save_strategy="epoch",
                         learning_rate = 2e-5,
                         num_train_epochs=3,
                         weight_decay=0.01)

## split

X=df_doc_encod[["sentence_id", "words"]]
Y=df_doc_encod["ner_tags"]
x_train, x_test, y_train, y_test = train_test_split(X,Y, test_size =0.2)
#building up train data and test data
train_data = pd.DataFrame({"sentence_id":x_train["sentence_id"],"words":x_train["words"],"labels":y_train})
test_data = pd.DataFrame({"sentence_id":x_test["sentence_id"],"words":x_test["words"],"labels":y_test})

## Metrics

from transformers import DataCollatorForTokenClassification
data_collator = DataCollatorForTokenClassification(tokenizer=tokenizer)

# !pip install seqeval
# !pip install evaluate

import evaluate
metric = evaluate.load('seqeval')

label_names=df_doc_encod['labels'].unique().tolist()

import numpy as np

def compute_metrics(eval_preds):
  logits, labels = eval_preds

  predictions = np.argmax(logits, axis=-1)

  true_labels = [[label_names[l] for l in label if l!=-100] for label in labels]

  true_predictions = [[label_names[p] for p,l in zip(prediction, label) if l!=-100]
                      for prediction, label in zip(predictions, labels)]

  all_metrics = metric.compute(predictions=true_predictions, references=true_labels)

  return {"precision": all_metrics['overall_precision'],
          "recall": all_metrics['overall_recall'],
          "f1": all_metrics['overall_f1'],
          "accuracy": all_metrics['overall_accuracy']}

from transformers import Trainer
trainer = Trainer(model=model,
                  args=args,
                  train_dataset = train_data,
                  eval_dataset = test_data,
                #   data_collator=data_collator,
                  compute_metrics=compute_metrics,
                  tokenizer=tokenizer)

trainer.train()

# 05_ random forest classifier

def feature_map(word):
    return np.array([word.istitle(), word.islower(), word.isupper(), len(word),
                     word.isdigit(),  word.isalpha()])

words = [feature_map(w) for w in df_doc_encod["words"].values.tolist()]
tags = df_doc_encod["labels"].values.tolist()

def feature_map(word):
    if isinstance(word, str):
        return np.array([word.istitle(), word.islower(), word.isupper(), len(word),
                         word.isdigit(), word.isalpha()])
    else:
        # Handle the case when the word is not a string (e.g., a float)
        return np.zeros(6)  # Return a default array of zeros for non-string values

words = [feature_map(w) for w in df_doc_encod["words"].values.tolist()]
tags = df_doc_encod["labels"].values.tolist()


words[:5]

import pandas as pd
import numpy as np
#Data visualisation
import matplotlib.pyplot as plt
import seaborn as sns
sns.set(color_codes=True)
sns.set(font_scale=1)
%matplotlib inline
%config InlineBackend.figure_format = 'svg'
#Modeling
from sklearn.model_selection import cross_val_predict, cross_val_score
from sklearn.ensemble import RandomForestClassifier

pred = cross_val_predict(RandomForestClassifier(n_estimators=20),X=words, y=tags, cv=5)

#Lets check the performance 
from sklearn.metrics import classification_report
report = classification_report(y_pred=pred, y_true=tags)
print(report)

# https://www.kaggle.com/code/bavalpreet26/ner-using-crf
